Concept-implementation of adding namespace while supporting a legacy code.
