+++ composer.json
@@ -33,6 +33,9 @@
         "squizlabs/php_codesniffer": "3.*"
     },
     "autoload": {
-        "classmap": ["lib/omise/"]
+        "psr-4": {
+            "Omise\\" : "lib/"
+        },
+        "classmap": [ "lib/omise/" ]
     }
 }
