+++ lib/Omise.php
@@ -10,6 +10,9 @@ require_once dirname(__FILE__).'/omise/Http/Response/Handler.php';
 require_once dirname(__FILE__).'/omise/exception/OmiseExceptions.php';
 
 // API Resources.
+require_once dirname(__FILE__).'/Account.php';
+
+// API Resources - Legacy classes.
 require_once dirname(__FILE__).'/omise/OmiseAccount.php';
 require_once dirname(__FILE__).'/omise/OmiseBalance.php';
 require_once dirname(__FILE__).'/omise/OmiseCard.php';
