+++ lib/omise/OmiseAccount.php
@@ -1,40 +1,8 @@
 <?php
-use Omise\Res\OmiseApiResource;
-
-class OmiseAccount extends OmiseApiResource
+/**
+ * @deprecated 3.0.0  No longer used this class and will be removed on Jan 4th, 2020.
+ */
+class OmiseAccount extends \Omise\Account
 {
-    const ENDPOINT = 'account';
 
-    /**
-     * Retrieves an account.
-     *
-     * @param  string $publickey
-     * @param  string $secretkey
-     *
-     * @return OmiseAccount
-     */
-    public static function retrieve($publickey = null, $secretkey = null)
-    {
-        return parent::g_retrieve(get_class(), self::getUrl(), $publickey, $secretkey);
-    }
-
-    /**
-     * (non-PHPdoc)
-     *
-     * @see OmiseApiResource::g_reload()
-     */
-    public function reload()
-    {
-        parent::g_reload(self::getUrl());
-    }
-
-    /**
-     * @param  string $id
-     *
-     * @return string
-     */
-    private static function getUrl($id = '')
-    {
-        return \Omise\ApiRequestor::OMISE_API_URL . self::ENDPOINT . '/' . $id;
-    }
 }
